[Security Solution] render notes new tools flyout in Security Solution and Discover
