+++ x-pack/solutions/security/plugins/security_solution/public/flyout/attack_details/left/tabs/notes_tab.tsx
@@ -5,29 +5,24 @@
  * 2.0.
  */
 
-import React, { memo } from 'react';
-import { EuiPanel } from '@elastic/eui';
-import { NOTES_TAB_CONTENT_TEST_ID } from '../../constants/test_ids';
+import React, { memo, useMemo } from 'react';
+import { buildDataTableRecord, type EsHitRecord } from '@kbn/discover-utils';
+import { NotesDetailsContent } from '../../../../flyout_v2/notes/components/notes_details_content';
+import { useTimelineConfig } from '../../../../flyout_v2/notes/hooks/use_timeline_config';
 import { useAttackDetailsContext } from '../../context';
-import { NotesDetailsContent } from '../../../shared/components/notes_details_content';
-import { useTimelineConfig } from '../../../document_details/shared/hooks/use_timeline_config';
+import { useWhichFlyout } from '../../../document_details/shared/hooks/use_which_flyout';
+import { Flyouts } from '../../../document_details/shared/constants/flyouts';
 
 /**
  * Notes tab content for the Attack Details flyout left panel.
  */
 export const NotesTab = memo(() => {
   const { attackId, searchHit } = useAttackDetailsContext();
-  const timelineConfig = useTimelineConfig(attackId);
+  const hit = useMemo(() => buildDataTableRecord(searchHit as unknown as EsHitRecord), [searchHit]);
+  const isTimelineFlyout = useWhichFlyout() === Flyouts.timeline;
+  const timelineConfig = useTimelineConfig(attackId, isTimelineFlyout);
 
-  return (
-    <EuiPanel data-test-subj={NOTES_TAB_CONTENT_TEST_ID} hasShadow={false}>
-      <NotesDetailsContent
-        documentId={attackId}
-        searchHit={searchHit}
-        timelineConfig={timelineConfig}
-      />
-    </EuiPanel>
-  );
+  return <NotesDetailsContent hit={hit} timelineConfig={timelineConfig} hideTimelineIcon={false} />;
 });
 
 NotesTab.displayName = 'NotesTab';
