+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/tabs/test_ids.ts
@@ -27,4 +27,3 @@ export const INSIGHTS_TAB_CORRELATIONS_BUTTON_TEST_ID =
   `${INSIGHTS_TAB_TEST_ID}CorrelationsButton` as const;
 export const INVESTIGATION_TAB_CONTENT_TEST_ID = `${PREFIX}InvestigationsTabContent` as const;
 export const RESPONSE_TAB_CONTENT_TEST_ID = `${PREFIX}ResponseTabContent` as const;
-export const NOTES_TAB_CONTENT_TEST_ID = `${PREFIX}NotesTabContent` as const;
