+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/notes/index.tsx
@@ -1,0 +1,54 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React, { memo } from 'react';
+import type { DataTableRecord } from '@kbn/discover-utils';
+import { EuiPanel } from '@elastic/eui';
+import { useTimelineConfig } from './hooks/use_timeline_config';
+import { useIsInSecurityApp } from '../../common/hooks/is_in_security_app';
+import {
+  FETCH_NOTES_ERROR,
+  NO_NOTES,
+  NotesDetailsContent,
+} from './components/notes_details_content';
+import { NOTES_DETAILS_TEST_ID } from './test_ids';
+
+export { FETCH_NOTES_ERROR, NO_NOTES };
+
+export interface NotesDetailsProps {
+  /**
+   * Document record used to fetch and associate notes and to derive the document type.
+   */
+  hit: DataTableRecord;
+  /**
+   * Whether the component is rendered inside the Timeline flyout.
+   * Controls whether the "attach to timeline" UI is shown.
+   */
+  isTimelineFlyout: boolean;
+}
+
+/**
+ * List all the notes for a document id and allows to create new notes associated with that document.
+ * Displayed in the document details expandable flyout left section.
+ */
+export const NotesDetails = memo(({ hit, isTimelineFlyout }: NotesDetailsProps) => {
+  const eventId = hit.raw._id ?? '';
+  const timelineConfig = useTimelineConfig(eventId, isTimelineFlyout);
+  const isInSecurityApp = useIsInSecurityApp();
+
+  return (
+    <EuiPanel data-test-subj={NOTES_DETAILS_TEST_ID} hasBorder={false} hasShadow={false}>
+      <NotesDetailsContent
+        hit={hit}
+        timelineConfig={timelineConfig}
+        hideTimelineIcon={!isInSecurityApp}
+      />
+    </EuiPanel>
+  );
+});
+
+NotesDetails.displayName = 'NotesDetails';
