+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/notes/test_ids.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { PREFIX } from '../../flyout/shared/test_ids';
+
+export const NOTES_DETAILS_TEST_ID = `${PREFIX}NotesTabContent` as const;
