+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/shared/components/flyout_provider.tsx
@@ -17,6 +17,9 @@ import { NavigationProvider } from '@kbn/security-solution-navigation';
 import type { StartServices } from '../../../types';
 import { ReactQueryClientProvider } from '../../../common/containers/query_client/query_client_provider';
 import { KibanaContextProvider } from '../../../common/lib/kibana';
+import { UserPrivilegesProvider } from '../../../common/components/user_privileges/user_privileges_context';
+import { UpsellingProvider } from '../../../common/components/upselling_provider';
+import { DiscoverInTimelineContextProvider } from '../../../common/components/discover_in_timeline/provider';
 
 export const flyoutProviders = ({
   services,
@@ -46,7 +49,15 @@ export const flyoutProviders = ({
       >
         <NavigationProvider core={services}>
           <Provider store={store}>
-            <ReactQueryClientProvider>{flyoutContent}</ReactQueryClientProvider>
+            <ReactQueryClientProvider>
+              <UserPrivilegesProvider kibanaCapabilities={services.application.capabilities}>
+                <UpsellingProvider upsellingService={services.upselling}>
+                  <DiscoverInTimelineContextProvider>
+                    {flyoutContent}
+                  </DiscoverInTimelineContextProvider>
+                </UpsellingProvider>
+              </UserPrivilegesProvider>
+            </ReactQueryClientProvider>
           </Provider>
         </NavigationProvider>
       </CellActionsProvider>
