+++ x-pack/solutions/security/plugins/security_solution/public/one_discover/alert_flyout_header_component/index.test.tsx
@@ -22,6 +22,16 @@ jest.mock('../../flyout_v2/document/header', () => ({
   Header: (props: unknown) => mockDocumentHeader(props),
 }));
 
+jest.mock('../../common/components/user_privileges/user_privileges_context', () => ({
+  UserPrivilegesProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
+}));
+
+jest.mock('../../common/components/discover_in_timeline/provider', () => ({
+  DiscoverInTimelineContextProvider: ({ children }: { children: React.ReactNode }) => (
+    <>{children}</>
+  ),
+}));
+
 describe('AlertFlyoutHeader', () => {
   beforeEach(() => {
     mockDocumentHeader.mockClear();
@@ -32,6 +42,12 @@ describe('AlertFlyoutHeader', () => {
     uiActions: {
       getTriggerCompatibleActions: jest.fn().mockResolvedValue([]),
     },
+    application: {
+      capabilities: {
+        securitySolution: { show: true, crud: true },
+      },
+    },
+    upselling: {},
   } as unknown as StartServices;
 
   it('wraps the header in KibanaContextProvider and ReactQueryClientProvider', async () => {
