+++ x-pack/solutions/security/plugins/security_solution/public/one_discover/alert_flyout_overview_tab_component/index.test.tsx
@@ -16,8 +16,18 @@ import { createStore } from 'redux';
 import { AlertFlyoutOverviewTab } from '.';
 import type { StartServices } from '../../types';
 
+jest.mock('../../common/components/user_privileges/user_privileges_context', () => ({
+  UserPrivilegesProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
+}));
+
 jest.mock('../../flyout_v2/document/tabs/overview_tab', () => ({
   OverviewTab: () => <div>{'MockOverviewTab'}</div>,
+}));
+
+jest.mock('../../common/components/discover_in_timeline/provider', () => ({
+  DiscoverInTimelineContextProvider: ({ children }: { children: React.ReactNode }) => (
+    <>{children}</>
+  ),
 }));
 
 const mockUseInitDataViewManager = jest.fn();
@@ -37,6 +47,12 @@ describe('AlertFlyoutOverviewTab', () => {
     uiActions: {
       getTriggerCompatibleActions: jest.fn().mockResolvedValue([]),
     },
+    application: {
+      capabilities: {
+        securitySolution: { show: true, crud: true },
+      },
+    },
+    upselling: {},
   } as unknown as StartServices;
 
   beforeEach(() => {
